:arrow_up: :lock: Upgrade debug package
Fixes [#5](https://github.com/twuni/ante-boilerplate-js/issues/5).

The **nyc** package has a transitive dependency on **debug@2.6.8**,
which we discovered via Snyk has a low-severity ReDoS security
vulnerability. This commit attempts to override this dependency
with a newer version in order to provide a short-term patch until
the **nyc** package is updated to no longer include this vulnerability.