:bug: Further attempts to get CircleCI test results
The documentation on requirements for having CircleCI's
`store_test_results` step actually pick up on standard
JUnit test reports is provided by-example. It seems to
imply that the only requirement is that the reports are
JUnit XML files, and that they reside within the directory
given to the `store_test_results` step. This does not
appear to be the case, as even when these two requirements
are met, the corresponding build has shown no test results.

This commit strictly adheres to the precise examples as given
in the documentation, in a last-ditch attempt to get test
results wired in. Additionally, it collects these test reports
as build artifacts for manual examination.