:no_entry_sign: Remove nyc due to unpatched security vulnerability
Istanbul is the current state-of-the-art in code coverage reporting
for JavaScript. However, there has been an unpatched security
vulnerability in its dependency graph for over two weeks already.
This, in addition to other factors around the project's health, have
contributed to this decision to abandon code coverage reporting
altogether, until the state-of-the-art comes to an acceptable level.